+++ src/Raven.CodeAnalysis/Compilation.Diagnostics.cs
@@ -48,9 +48,6 @@ public partial class Compilation
         if (Options.AllowUnsafe)
             Add(Diagnostic.Create(CompilerDiagnostics.UnsafeModeEnabled, Location.None));
 
-        if (Options.MembersPublicByDefaultConfigured && !Options.MembersPublicByDefault)
-            Add(Diagnostic.Create(CompilerDiagnostics.ExplicitPublicAccessibilityRequired, Location.None));
-
         return diagnostics.OrderBy(static diagnostic => diagnostic, DiagnosticComparer.Instance).ToImmutableArray();
 
         void AddTreeDiagnostics(SyntaxTree syntaxTree)